test(mcp): split OAuth CLI coverage
