+++ src/cli/mcp-cli.test.ts
@@ -1,124 +1,32 @@
 // MCP CLI tests cover MCP command registration and server configuration behavior.
 import fs from "node:fs/promises";
-import os from "node:os";
 import path from "node:path";
-import { Command } from "commander";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import * as mcpHttpFetch from "../agents/mcp-http-fetch.js";
 import { withTempHome } from "../config/home-env.test-harness.js";
 import { createDeferred } from "../shared/deferred.js";
-import { getFreePort } from "../test-utils/ports.js";
-import { registerMcpCli } from "./mcp-cli.js";
 import { writeProbeMcpServer } from "./mcp-cli.test-support.js";
-
-type CreateSessionMcpRuntime =
-  typeof import("../agents/agent-bundle-mcp-runtime.js").createSessionMcpRuntime;
-type RunMcpOAuthLogin = typeof import("../agents/mcp-oauth.js").runMcpOAuthLogin;
-
-const mocks = vi.hoisted(() => {
-  const runtime = {
-    log: vi.fn(),
-    error: vi.fn(),
-    exit: vi.fn((code: number) => {
-      throw new Error(`__exit__:${code}`);
-    }),
-    writeJson: vi.fn((value: unknown, space = 2) => {
-      runtime.log(JSON.stringify(value, null, space > 0 ? space : undefined));
-    }),
-  };
-  return {
-    runtime,
-    serveOpenClawChannelMcp: vi.fn(),
-    clearMcpOAuthCredentials: vi.fn(),
-    readMcpOAuthCredentialsStatus: vi.fn(),
-    runMcpOAuthLogin: vi.fn(),
-    createSessionMcpRuntimeOverride: undefined as CreateSessionMcpRuntime | undefined,
-  };
-});
-
-const defaultRuntime = mocks.runtime;
-const mockLog = defaultRuntime.log;
-const mockError = defaultRuntime.error;
-const serveOpenClawChannelMcp = mocks.serveOpenClawChannelMcp;
-const clearMcpOAuthCredentials = mocks.clearMcpOAuthCredentials;
-const readMcpOAuthCredentialsStatus = mocks.readMcpOAuthCredentialsStatus;
-const runMcpOAuthLogin = mocks.runMcpOAuthLogin;
-
-vi.mock("../runtime.js", () => ({
-  defaultRuntime: mocks.runtime,
-}));
-
-vi.mock("../mcp/channel-server.js", () => ({
-  serveOpenClawChannelMcp: mocks.serveOpenClawChannelMcp,
-}));
-
-vi.mock("../agents/mcp-oauth.js", () => ({
-  clearMcpOAuthCredentials: mocks.clearMcpOAuthCredentials,
-  readMcpOAuthCredentialsStatus: mocks.readMcpOAuthCredentialsStatus,
-  runMcpOAuthLogin: mocks.runMcpOAuthLogin,
-}));
-
-vi.mock("../agents/agent-bundle-mcp-runtime.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../agents/agent-bundle-mcp-runtime.js")>();
-  return {
-    ...actual,
-    createSessionMcpRuntime: (params: Parameters<CreateSessionMcpRuntime>[0]) =>
-      mocks.createSessionMcpRuntimeOverride?.(params) ?? actual.createSessionMcpRuntime(params),
-  };
-});
-
-const tempDirs: string[] = [];
-
-async function createWorkspace(): Promise<string> {
-  const dir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-cli-mcp-"));
-  tempDirs.push(dir);
-  return dir;
-}
-
-let sharedProgram: Command;
-
-async function runMcpCommand(args: string[]) {
-  await sharedProgram.parseAsync(args, { from: "user" });
-}
-
-function lastLogLine(): string {
-  return lastRuntimeLine(mockLog);
-}
-
-function lastErrorLine(): string {
-  return lastRuntimeLine(mockError);
-}
-
-function lastRuntimeLine(mock: typeof mockLog): string {
-  const call = mock.mock.calls[mock.mock.calls.length - 1];
-  return String(call?.[0] ?? "");
-}
+import {
+  cleanupMcpCliTestState,
+  clearMcpOAuthCredentials,
+  createWorkspace,
+  lastErrorLine,
+  lastLogLine,
+  mockError,
+  mockLog,
+  readMcpOAuthCredentialsStatus,
+  resetMcpCliTestState,
+  runMcpCommand,
+  setCreateSessionMcpRuntimeOverride,
+  serveOpenClawChannelMcp,
+} from "./mcp-cli.test-harness.js";
 
 describe("mcp cli", () => {
-  if (!sharedProgram) {
-    sharedProgram = new Command();
-    sharedProgram.exitOverride();
-    registerMcpCli(sharedProgram);
-  }
-
   beforeEach(() => {
-    vi.clearAllMocks();
-    mocks.createSessionMcpRuntimeOverride = undefined;
-    readMcpOAuthCredentialsStatus.mockResolvedValue({
-      hasTokens: false,
-      requiresAuthorization: false,
-      hasClientInformation: false,
-      hasCodeVerifier: false,
-      hasDiscoveryState: false,
-      hasLastAuthorizationUrl: false,
-    });
+    resetMcpCliTestState();
   });
 
   afterEach(async () => {
-    vi.restoreAllMocks();
-    await Promise.all(
-      tempDirs.splice(0).map((dir) => fs.rm(dir, { recursive: true, force: true })),
-    );
+    await cleanupMcpCliTestState();
   });
 
   it("sets and shows a configured MCP server", async () => {
@@ -301,7 +209,7 @@ describe("mcp cli", () => {
       const configPath = path.join(home, ".openclaw", "openclaw.json");
       vi.spyOn(process, "cwd").mockReturnValue(workspaceDir);
       let probeTimeoutMs: unknown;
-      mocks.createSessionMcpRuntimeOverride = (params) => {
+      setCreateSessionMcpRuntimeOverride((params) => {
         probeTimeoutMs = params.cfg?.mcp?.servers?.["hung-default"]?.connectionTimeoutMs;
         return {
           sessionId: params.sessionId,
@@ -329,7 +237,7 @@ describe("mcp cli", () => {
           callTool: async () => ({ content: [] }),
           dispose: async () => {},
         };
-      };
+      });
 
       await expect(
         runMcpCommand(["mcp", "add", "hung-default", "--command", process.execPath]),
@@ -474,232 +382,6 @@ describe("mcp cli", () => {
       expect(verboseOutput).toContain("launch: node server.mjs --api-key *** --token=***");
       expect(verboseOutput).not.toContain("test-api-key");
       expect(verboseOutput).not.toContain("test-token");
-    });
-  });
-
-  it("includes OAuth credential status in MCP status output", async () => {
-    await withTempHome("openclaw-cli-mcp-home-", async () => {
-      const workspaceDir = await createWorkspace();
-      vi.spyOn(process, "cwd").mockReturnValue(workspaceDir);
-      readMcpOAuthCredentialsStatus.mockResolvedValueOnce({
-        hasTokens: true,
-        requiresAuthorization: false,
-        hasClientInformation: true,
-        hasCodeVerifier: false,
-        hasDiscoveryState: true,
-        hasLastAuthorizationUrl: true,
-      });
-
-      await runMcpCommand([
-        "mcp",
-        "set",
-        "docs",
-        '{"url":"https://mcp.example.com","transport":"streamable-http","auth":"oauth"}',
-      ]);
-      mockLog.mockClear();
-
-      await runMcpCommand(["mcp", "status", "--json"]);
-
-      expect(JSON.parse(lastLogLine()).servers[0]).toMatchObject({
-        name: "docs",
-        auth: "oauth",
-        authStatus: {
-          hasTokens: true,
-          requiresAuthorization: false,
-          hasClientInformation: true,
-          hasCodeVerifier: false,
-          hasDiscoveryState: true,
-          hasLastAuthorizationUrl: true,
-        },
-      });
-    });
-  });
-
-  it("surfaces required OAuth authorization in status and doctor", async () => {
-    await withTempHome("openclaw-cli-mcp-home-", async () => {
-      const workspaceDir = await createWorkspace();
-      vi.spyOn(process, "cwd").mockReturnValue(workspaceDir);
-      readMcpOAuthCredentialsStatus.mockResolvedValue({
-        hasTokens: true,
-        requiresAuthorization: true,
-        hasClientInformation: true,
-        hasCodeVerifier: false,
-        hasDiscoveryState: true,
-        hasLastAuthorizationUrl: true,
-      });
-
-      await runMcpCommand([
-        "mcp",
-        "set",
-        "docs",
-        '{"url":"https://mcp.example.com","transport":"streamable-http","auth":"oauth"}',
-      ]);
-      mockLog.mockClear();
-
-      await runMcpCommand(["mcp", "status", "--verbose"]);
-
-      const statusLines = mockLog.mock.calls.map((call) => String(call[0]));
-      expect(statusLines).toContain("- docs: streamable-http oauth authorization-required");
-      expect(statusLines).toContain("  oauth: tokens=yes authorization=required client=yes");
-
-      mockLog.mockClear();
-      await runMcpCommand(["mcp", "doctor", "--json"]);
-
-      expect(JSON.parse(lastLogLine())).toMatchObject({
-        ok: true,
-        servers: [
-          {
-            name: "docs",
-            ok: true,
-            issues: [
-              {
-                level: "warning",
-                message:
-                  "OAuth credentials require additional authorization; run openclaw mcp login docs",
-              },
-            ],
-          },
-        ],
-      });
-    });
-  });
-
-  it("configures enablement, timeouts, and OAuth login", async () => {
-    await withTempHome("openclaw-cli-mcp-home-", async () => {
-      const workspaceDir = await createWorkspace();
-      vi.spyOn(process, "cwd").mockReturnValue(workspaceDir);
-      const buildMcpHttpFetch = vi.spyOn(mcpHttpFetch, "buildMcpHttpFetch");
-      runMcpOAuthLogin.mockResolvedValueOnce("authorized");
-
-      await runMcpCommand([
-        "mcp",
-        "set",
-        "docs",
-        '{"url":"https://mcp.example.com","transport":"streamable-http"}',
-      ]);
-      await runMcpCommand([
-        "mcp",
-        "configure",
-        "docs",
-        "--disable",
-        "--timeout",
-        "9",
-        "--auth",
-        "oauth",
-      ]);
-      await runMcpCommand(["mcp", "login", "docs", "--code", "abc123"]);
-
-      expect(buildMcpHttpFetch).toHaveBeenCalledWith(
-        expect.objectContaining({
-          resourceUrl: "https://mcp.example.com",
-          timeoutMs: 9_000,
-        }),
-      );
-      expect(runMcpOAuthLogin).toHaveBeenCalledWith({
-        serverName: "docs",
-        serverUrl: "https://mcp.example.com",
-        config: undefined,
-        fetchFn: expect.any(Function),
-        authorizationCode: "abc123",
-        onAuthorizationUrl: expect.any(Function),
-      });
-
-      mockLog.mockClear();
-      await runMcpCommand(["mcp", "status", "--json"]);
-      expect(JSON.parse(lastLogLine()).servers[0]).toMatchObject({
-        name: "docs",
-        enabled: false,
-        ok: false,
-        requestTimeoutMs: 9_000,
-        auth: "oauth",
-      });
-    });
-  });
-
-  it("captures the browser redirect before completing OAuth login", async () => {
-    await withTempHome("openclaw-cli-mcp-home-", async () => {
-      const workspaceDir = await createWorkspace();
-      const port = await getFreePort();
-      const redirectUrl = `http://127.0.0.1:${port}/oauth/callback`;
-      const authUrl = new URL("https://auth.example.com/authorize");
-      authUrl.searchParams.set("state", "state-1");
-      authUrl.searchParams.set("redirect_uri", redirectUrl);
-      vi.spyOn(process, "cwd").mockReturnValue(workspaceDir);
-
-      await runMcpCommand([
-        "mcp",
-        "set",
-        "docs",
-        JSON.stringify({
-          url: "https://mcp.example.com",
-          transport: "streamable-http",
-          auth: "oauth",
-          oauth: { redirectUrl },
-        }),
-      ]);
-      mockLog.mockClear();
-      let capturedCode: string | undefined;
-      runMcpOAuthLogin.mockImplementationOnce(async (params: Parameters<RunMcpOAuthLogin>[0]) => {
-        const code = await params.onAuthorizationUrl?.(authUrl);
-        capturedCode = typeof code === "string" ? code : undefined;
-        return "authorized";
-      });
-
-      const login = runMcpCommand(["mcp", "login", "docs"]);
-      await vi.waitFor(() => expect(mockLog).toHaveBeenCalledWith(authUrl.toString()));
-      const response = await fetch(`${redirectUrl}?code=browser-code&state=state-1`);
-      expect(response.status).toBe(200);
-      await expect(response.text()).resolves.toContain("MCP OAuth complete");
-      await login;
-
-      expect(capturedCode).toBe("browser-code");
-      expect(lastLogLine()).toBe('MCP OAuth credentials saved for "docs".');
-      expect(mockLog).toHaveBeenCalledWith(
-        "If the browser redirect cannot reach this machine, stop this command and run openclaw mcp login docs --code <code>.",
-      );
-    });
-  });
-
-  it("clears stored OAuth credentials on logout", async () => {
-    await withTempHome("openclaw-cli-mcp-home-", async () => {
-      const workspaceDir = await createWorkspace();
-      vi.spyOn(process, "cwd").mockReturnValue(workspaceDir);
-
-      await runMcpCommand([
-        "mcp",
-        "set",
-        "docs",
-        '{"url":"https://mcp.example.com","transport":"streamable-http","auth":"oauth"}',
-      ]);
-      clearMcpOAuthCredentials.mockClear();
-      await runMcpCommand(["mcp", "logout", "docs"]);
-
-      expect(clearMcpOAuthCredentials).toHaveBeenCalledWith({
-        serverName: "docs",
-        serverUrl: "https://mcp.example.com",
-      });
-      expect(lastLogLine()).toBe('MCP OAuth credentials cleared for "docs".');
-    });
-  });
-
-  it("clears stored OAuth credentials after auth is removed", async () => {
-    await withTempHome("openclaw-cli-mcp-home-", async () => {
-      const workspaceDir = await createWorkspace();
-      vi.spyOn(process, "cwd").mockReturnValue(workspaceDir);
-
-      await runMcpCommand([
-        "mcp",
-        "set",
-        "docs",
-        '{"url":"https://mcp.example.com","transport":"streamable-http"}',
-      ]);
-      clearMcpOAuthCredentials.mockClear();
-      await runMcpCommand(["mcp", "logout", "docs"]);
-
-      expect(clearMcpOAuthCredentials).toHaveBeenCalledWith({
-        serverName: "docs",
-        serverUrl: "https://mcp.example.com",
-      });
     });
   });
